feat: use object for management details in session endpoint
