+++ frontend/src/components/header.tsx
@@ -217,7 +217,7 @@ function ClassGroupManagerMenu() {
 
   const isSystemAdmin = session.data?.user.role == UserRole.SystemAdmin;
   const isClassGroupManager =
-    session.data?.has_managed_class_groups || isSystemAdmin;
+    session.data?.management_details.has_managed_class_groups || isSystemAdmin;
 
   if (!isClassGroupManager) {
     return null;
